+++ packages/browser-wallet/CHANGELOG.md
@@ -11,6 +11,8 @@
 -   The About page link to the terms and conditions pointed to the wrong URL. It now uses the value retrieved from the wallet proxy, or the correct default to the unified terms and conditions page.
 -   Link to transaction log on stagenet was incorrect.
 -   Fixed an empty recovery displaying an error instead of informing the user that nothing was found.
+-   An issue where the transaction list view would show the `Request CCD` button while loading the initial batch of transactions.
+-   Added a missing translation for the `Request CCD` button.
 
 ## 1.0.4