+++ packages/browser-wallet/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 -   A build script for making builds that include the stagenet network.
 -   `deployModule` transactions now display the hash instead of the entire module source.
+-   Text that a transaction has been submitted.
 
 ### Fixed
 
@@ -15,6 +16,10 @@
 -   An issue where the transaction list view would show the `Request CCD` button while loading the initial batch of transactions.
 -   Added a missing translation for the `Request CCD` button.
 -   `deployModule` transactions no longer cause the wallet to crash, when supplied with `sendTransaction`.
+
+### Changed
+
+-   Messages when confirming baker/delegation transactions no longer appears after the transaction has been submitted.
 
 ## 1.0.4